add delete ques

+++ src/pages/MainPage.tsx
@@ -61,32 +61,29 @@ function MainPage() {
 
   async function deleteTask(dealId: string) {
     try {
-      const res = await fetch(`${API_URL}/${dealId}`, {
-        method: 'DELETE',
-      });
-      if (!res.ok) {
-        throw new Error('Network response was not ok');
-      }
+      if (confirm('Точно хочешь удалить?')) {
+        const res = await fetch(`${API_URL}/${dealId}`, {
+          method: 'DELETE',
+        });
+        if (!res.ok) {
+          throw new Error('Network response was not ok');
+        }
 
-      setDeals((prevDeals) => prevDeals.filter((deal) => deal.id !== Number(dealId)));
+        setDeals((prevDeals) => prevDeals.filter((deal) => deal.id !== Number(dealId)));
+      }
+      return
     } catch (error) {
       console.error('Error deleting deal:', error);
     }
   }
   return (
     <>
-      <h1 className='title' >Todo List</h1>
+      <h1 className='title'>Todo List</h1>
       <div className='links'>
-      <NavLink 
-          to='/' 
-          className={({ isActive }) => (isActive ? 'active-link' : 'link')}
-        >
+        <NavLink to='/' className={({ isActive }) => (isActive ? 'active-link' : 'link')}>
           in progress
         </NavLink>
-        <NavLink 
-          to='/progress' 
-          className={({ isActive }) => (isActive ? 'active-link' : 'link')} 
-        >
+        <NavLink to='/progress' className={({ isActive }) => (isActive ? 'active-link' : 'link')}>
           is done
         </NavLink>
       </div>
@@ -103,7 +100,7 @@ function MainPage() {
       {deals?.some((deal) => !deal.isActive) ? (
         <List deals={deals.filter((deal) => !deal.isActive)} onDelete={deleteTask} />
       ) : (
-        <h1  className='empty'>Список пуст :(</h1>
+        <h1 className='empty'>Список пуст :(</h1>
       )}
     </>
   );
